add cross-platform AutoBatchSize (#3143)
See https://github.com/deepmodeling/deepmd-kit/issues/3118 and
https://github.com/dptech-corp/deepmd-pytorch/issues/137. Subclass needs
to implement `is_gpu_available` and `is_oom_error`.
